+++ src/console.h
@@ -40,7 +40,7 @@ extern INT32 con_destlines;
 
 extern INT32 con_clearlines; // lines of top of screen to refresh
 extern boolean con_hudupdate; // hud messages have changed, need refresh
-extern size_t con_scalefactor; // console text scale factor
+extern UINT32 con_scalefactor; // console text scale factor
 
 extern consvar_t cons_backcolor;
